Remove computerSelection argument
Now the variable that stores the computer decision is inside the
playRound function. This way, the decision the computer takes changes
every time a round of the game is played.
Since there's no need for the computerSelection argument anymore,
it was deleted and now the playRound function only needs one argument.
This one argument being the choice of the user.